Enhance CustomInstaller builder to support archives and local data in addition to git URI's
Required significant refactoring of the design of the CustomInstaller builder - It now always creates a single `.tar.gz` file per custom-installer that includes the "package" (whatever the URI / local data points to) + the install script, so there's always just one file per custom installer to transfer into the image for installing.

This has several upsides:

- Get rid of `.git` directory in the image.
- Since the `.tar.gz` file is extracted & installed & deleted in the same Dockerfile layer - the layer size is kept minimal - at the size of the tarball.
- A preparation step for a next step in which the tarball isn't even copied into the image, but served by a YABT-spawned server during build-time and deleted within the same Docekrfile layer, so the layer size is minimized to *zero* (or just the actual installed artifacts).

The archive support was needed for protobuf installation, and the local-data support was needed for spark-conf "installation". Git support exists for rpcz installation, but not yet used (since it depends on protobuf and other stuff).