Implement DockerImage target caching using distributed image registry without assuming shared Docker build cache
The default behavior should remain unchanged when building Docker images.

The new behavior is controlled by a new argument to the DockerImage builder - a `image_caching_behavior` dictionary.

When specified, the dictionary can be used to control:

1. Whether an image is pulled from a image registry before building it (using a potentially different remote name & tag).
    - If a local image is not cached.
    - If a local image is cached.
2. If a local image is cached (whether it was just pulled or not) - whether to run a build, or skip it and just use the cache instead (and tag the remote as the local, if the names are different).
3. If a local image is not cached - whether to build it or fail (to allow more control over building of base images).
4. If a build completed successfully - whether to push it to an image registry (using a potentially different remote name & tag).

Several command-line flags were added to allow YBuild's to provide "dynamic" values for these dictionary fields (e.g. "pull when not offline, build when flag `--build-base-images` flag is given").

See example usage in the parallel PR on yowza1 that adds this dictionary to the ywz-run-base target.